Start decoupling socket/reducer, async-ify /auth
Decouple the socket and reducer as much as possible
by having the socket trigger a CustomEvent that dispatches
the data to the reducer rather than requiring the socket
to have a reference to the reducer (ew).

Use `async/await` to create /auth endpoint rather than
generators given that future versions of koa won't support
generators.